+++ dss/util/bundles.py
@@ -4,7 +4,7 @@ import typing
 from cloud_blobstore import BlobNotFoundError
 
 from dss import Config, DSSException, Replica
-from dss.hcablobstore import BundleFileMetadata, BundleMetadata
+from dss.storage.hcablobstore import BundleFileMetadata, BundleMetadata
 from dss.storage.bundles import DSS_BUNDLE_KEY_REGEX, DSS_BUNDLE_TOMBSTONE_REGEX, TombstoneID, BundleFQID
 from dss.util import UrlBuilder
 from dss.util.blobstore import test_object_exists
