+++ tests/test_gshcablobstore.py
@@ -10,7 +10,7 @@ from cloud_blobstore.gs import GSBlobStore
 pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))  # noqa
 sys.path.insert(0, pkg_root)  # noqa
 
-from dss.hcablobstore.gs import GSHCABlobStore
+from dss.storage.hcablobstore.gs import GSHCABlobStore
 from tests import infra
 from tests.hcablobstore_base import HCABlobStoreTests
